chore(sequelizer): reorganize structur folder
