+++ packages/user-acount/src/drivers/postgres/sequelizer.js
@@ -1,0 +1,17 @@
+const { Sequelize } = require('sequelize');
+
+const { config } = require('../../../config/config');
+const setupModels = require('../../domains/db/models');
+
+const USER = encodeURIComponent(config.dbUser);
+const PASSWORD = encodeURIComponent(config.dbPassword);
+const URI = `postgres://${USER}:${PASSWORD}@${config.dbHost}:${config.dbPort}/${config.dbName}`;
+
+const sequelize = new Sequelize(URI, {
+  dialect: 'postgres',
+  logging: false,
+});
+
+setupModels(sequelize);
+
+module.exports = sequelize;
